requirements.txt
flask>=2.0
flask-cors
click>=8.0
sentence-transformers>=2.2.2
faiss-cpu>=1.7.2
numpy>=1.19.0,<2
datasets>=2.0.0
tqdm>=4.0.0